+++ package/smooth_transition/CHANGELOG.md
@@ -4,3 +4,13 @@
 
 ## 1.0.1
 - Fix Readme code
+
+## 1.0.2
+- Improved performance of `SmoothTransition`.
+- Added support for reverse animation.
+- Fixed bug with alignment in `Stack`.
+- Updated example app with new usage.
+
+## 1.0.3
+- Fixed alert 
+- Updated example app with new usage.